[Workflow Mutation Status](4) Share action graph snapshot builder
